refactor: add computer save method
