+++ Database/DatabaseSetup.cs
@@ -10,7 +10,6 @@ class DatabaseSetup
     {
         _databaseConfig = databaseConfig;
         CreateComputerTable();
-        CreateLabTable();
     }
 
     private void CreateComputerTable()
@@ -30,10 +29,5 @@ class DatabaseSetup
 
         command.ExecuteNonQuery();
         connection.Close();
-    }
-
-    private void CreateLabTable()
-    {
-
     }
 }
